add jobID to global env, add adapter.go init
