+++ HadesLogManager/main.go
@@ -86,12 +86,7 @@ func connectNATS(config utils.NatsConfig) (*nats.Conn, error) {
 }
 
 // runWithGracefulShutdown starts services and handles graceful shutdown
-func runWithGracefulShutdown(
-	ctx context.Context,
-	cancel context.CancelFunc,
-	cfg HadesLogManagerConfig,
-	dynamicManager LogManager,
-	logAggregator LogAggregator,
+func runWithGracefulShutdown(ctx context.Context, cancel context.CancelFunc, cfg HadesLogManagerConfig, dynamicManager LogManager, logAggregator LogAggregator,
 ) error {
 	var wg sync.WaitGroup
 	errChan := make(chan error, 2)
@@ -134,12 +129,7 @@ func runWithGracefulShutdown(
 }
 
 // waitForShutdown waits for OS signal or error and performs graceful shutdown
-func waitForShutdown(
-	ctx context.Context,
-	cancel context.CancelFunc,
-	server *http.Server,
-	wg *sync.WaitGroup,
-	errChan chan error,
+func waitForShutdown(ctx context.Context, cancel context.CancelFunc, server *http.Server, wg *sync.WaitGroup, errChan chan error,
 ) error {
 	sigChan := make(chan os.Signal, 1)
 	signal.Notify(sigChan, syscall.SIGINT, syscall.SIGTERM)
