+++ docs/api/Create Build Job (Test Succeed) with Result Parse Step.bru
@@ -74,8 +74,7 @@ body:json {
         ],
         "metadata": {
           "API_ENDPOINT": "http://host.docker.internal:8082/adapter/test-results",
-          "INGEST_DIR": "/shared/example/build/test-results/test",
-          "JOB_NAME": "{jobId}"
+          "INGEST_DIR": "/shared/example/build/test-results/test"
         }
       }
     ]
